Replaced vuew FileUpload component with alpine js
